+++ plugins/09_Shadowlands/localization/zhTW.lua
@@ -353,6 +353,8 @@ L["sinrunner_note_day6"] = "從城牆破洞的 {npc:167815} 購買3個 {item:179
 
 L["options_icons_carriages"] = "馬車"
 L["options_icons_carriages_desc"] = "顯示可騎乘馬車的位置和路徑。"
+L["options_icons_dredbats"] = nil
+L["options_icons_dredbats_desc"] = nil
 L["options_icons_sinrunners"] = nil
 L["options_icons_sinrunners_desc"] = nil