+++ plugins/09_Shadowlands/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added Ve'nari to the map to track Maw and Torghast upgrades.
 * Added Family Exorcist achievements to the pet trainer tooltips.
 * Added Castle Sinrunner routes for the What We Ride in the Shadows achievement.
+* Added Dredbat Statue routes for the Bat! achievement.
 
 # v17